@@ -4,11 +4,12 @@
   "description": "Flexible Tree layout supporting ancestors, descendants and side nodes in all 4 orientations",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
+  "type": "module",
   "scripts": {
-    "browser": "open playground/index.html && watchify playground/source.js -p [ tsify ] -o playground/bundle.js",
-    "build": "tsc",
-    "dev": "tsc --watch",
-    "prepublishOnly": "yarn build",
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "prepublishOnly": "pnpm build",
     "test": "jest"
   },
   "files": [
@@ -19,18 +20,13 @@
     "type": "git",
     "url": "https://github.com/codeledge/entitree-flex"
   },
-  "dependencies": {},
   "devDependencies": {
-    "@babel/core": "^7.17.9",
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-typescript": "^7.16.7",
     "@svgdotjs/svg.js": "^3.1.2",
-    "@types/jest": "^27.4.1",
-    "babel-jest": "^28.0.0",
-    "jest": "^28.0.0",
-    "tsify": "^5.0.4",
-    "typescript": "^4.6.3",
-    "watchify": "^4.0.0"
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.14.0",
+    "jest": "^29.7.0",
+    "typescript": "~5.8.3",
+    "vite": "^6.2.0"
   },
   "keywords": [
     "tree",
@@ -42,5 +38,6 @@
     "parents",
     "family",
     "pedigree"
-  ]
+  ],
+  "packageManager": "pnpm@10.8.0+sha512.0e82714d1b5b43c74610193cb20734897c1d00de89d0e18420aebc5977fa13d780a9cb05734624e81ebd81cc876cd464794850641c48b9544326b5622ca29971"
 }
